chore: reorganize files, add new route welcome and compound component
