+++ src/components/App.test.tsx
@@ -5,5 +5,5 @@ import App from './App';
 test('initial site', () => {
   render(<App />);
 
-  expect(screen.getByText('Compound components in React')).toBeInTheDocument();
+  expect(screen.getByText('WIP: Welcome to the Reference App!')).toBeInTheDocument();
 });
