+++ src/components/App.tsx
@@ -1,13 +1,11 @@
 import '@code-hike/mdx/styles';
-import './css/custom-ch.css';
-import CompoundComponentsApp from './components/compoundComponents/index.mdx';
+import '../css/custom-ch.css';
 
 function App() {
   return (
     <article className='pl-[--page-side-nav-width] pt-8 md:pl-[--page-side-nav-width-md] xl:md:pl-[--page-side-nav-width-xl]'>
-      <div>
-        <CompoundComponentsApp />
-      </div>
+        WIP:
+        Welcome to the Reference App!
     </article>
   );
 }
