+++ src/components/SideNavigation.tsx
@@ -9,11 +9,11 @@ export const SideNavigation = () => {
   const navItems: NavItem[] = [
     {
       linkTo: '/',
-      name: 'Home',
+      name: 'Welcome',
     },
     {
-      linkTo: '/about',
-      name: 'About',
+      linkTo: '/compoundComponent',
+      name: 'Compound Component',
     },
   ];
 
@@ -24,19 +24,19 @@ export const SideNavigation = () => {
       aria-label="Sidebar"
     >
       <div className="h-full overflow-y-auto bg-gray-50 py-4">
-            <ul className="list-none space-y-2 border-l border-gray-600 p-0 ">
-              {navItems.map((item: NavItem) => (
-                  <li key={item.linkTo}>
-                    <Link
-                        to={item.linkTo}
-                        className="group flex items-center rounded-lg p-2 text-xl text-gray-900 hover:bg-gray-100 dark:hover:bg-gray-500 dark:hover:text-white [&.active]:font-bold"
-                    >
-                      <span className="ms-[--page-side-nav-padding] md:ms-[--page-side-nav-padding-md] xl:ms-[--page-side-nav-padding-xl]">{item.name}</span>
-                    </Link>
-                  </li>
-              ))}
-            </ul>
-          </div>
+        <ul className="list-none space-y-2 border-l border-gray-600 p-0 ">
+          {navItems.map((item: NavItem) => (
+              <li key={item.linkTo}>
+                <Link
+                    to={item.linkTo}
+                    className="group flex items-center rounded-lg p-2 text-xl text-gray-900 hover:bg-gray-100 dark:hover:bg-gray-500 dark:hover:text-white [&.active]:font-bold"
+                >
+                  <span className="ms-[--page-side-nav-padding] md:ms-[--page-side-nav-padding-md] xl:ms-[--page-side-nav-padding-xl]">{item.name}</span>
+                </Link>
+              </li>
+          ))}
+        </ul>
+      </div>
     </aside>
 );
 };
