+++ src/routes/index.lazy.tsx
@@ -1,5 +1,5 @@
 import { createLazyFileRoute } from '@tanstack/react-router';
-import App from '../App';
+import App from '../components/App';
 
 export const Route = createLazyFileRoute('/')({
   component: Index,
